% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffs.R
\name{cutoffs}
\alias{cutoffs}
\title{Creates NOT DONE, purpose of function}
\usage{
cutoffs(X, Y, C, c.vec, kk, cost, K = 20, Lip_0temp, Lip_1temp, B.0m, B.1m)
}
\arguments{
\item{X}{X value}

\item{Y}{Y value}

\item{C}{C value}

\item{c.vec}{A vector of cutoff values}

\item{kk}{Multiplicative factor}

\item{cost}{cost}

\item{K}{K-folds, default is 20}

\item{Lip_0temp}{Smoothness}

\item{Lip_1temp}{Smoothness}

\item{B.0m}{B0}

\item{B.1m}{B1}
}
\value{
NOT DONE: Smoothness parameters
}
\description{
Creates NOT DONE, purpose of function
}
\examples{
smooth(NOT DONE)

}
