% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoffsviz.R
\name{cutoffsviz}
\alias{cutoffsviz}
\title{Creates visualiation of new and old cutoff values based on output of optimalcutoffs()}
\usage{
cutoffsviz(data, color)
}
\arguments{
\item{data}{A data frame with the following values: Original Cutoffs, New Cutoffs, & Difference}

\item{color}{A palette of 2 colors}
}
\value{
Graph of new and old cutoff values
}
\description{
Creates visualiation of new and old cutoff values based on output of optimalcutoffs()
}
\examples{
cutoffsviz(NOT DONE)

}
