% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{lip_extra}
\alias{lip_extra}
\title{Creates NOT DONE, purpose of function e.g. extrapolation function}
\usage{
lip_extra(
  x.train,
  group,
  g,
  g.prim,
  Lip_1_initial,
  Lip_0_initial,
  B.1m_initial,
  B.0m_initial
)
}
\arguments{
\item{x.train}{...}

\item{group}{...}

\item{g}{...}

\item{g.prim}{...}

\item{Lip_1_initial}{...}

\item{Lip_0_initial}{...}

\item{B.1m_initial}{...}

\item{B.0m_initial}{...}
}
\value{
NOT DONE
}
\description{
Creates NOT DONE, purpose of function e.g. extrapolation function
}
\examples{
lip_extra(NOT DONE)

}
