% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimal.R
\name{optimal}
\alias{optimal}
\title{Creates NOT DONE, purpose of function}
\usage{
optimal(data_all, c.vec, kk = 1, Lip_1, Lip_0)
}
\arguments{
\item{data_all}{A data set with variables X, Y, C, W, G, and pseudo outcomes}

\item{c.vec}{A vector of cutoff values}

\item{kk}{Multiplicative factor on the smoothness parameter, set cost=0 when varying kk}

\item{Lip_1}{Initial value for smoothness parameter}

\item{Lip_0}{Initial value for smoothness parameter}
}
\value{
NOT DONE
}
\description{
Creates NOT DONE, purpose of function
}
\examples{
optimal(NOT DONE)

}
