% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalcutoffs.R
\name{optimalcutoffs}
\alias{optimalcutoffs}
\title{Creates NOT DONE, purpose of function}
\usage{
optimalcutoffs(X, Y, C, c.vec, K = 20, kk = 1, cost = NULL)
}
\arguments{
\item{X}{Uni-variate running variable}

\item{Y}{Outcome variable}

\item{C}{A vector of cutoff values}

\item{c.vec}{A vector of BASELINE cutoff values}

\item{K}{Folds}

\item{kk}{Multiplicative factor on the smoothness parameter, set cost=0 when varying kk}

\item{cost}{Cost value}
}
\value{
NOT DONE
}
\description{
Creates NOT DONE, purpose of function
}
\examples{
optimalcutoffs(NOT DONE)

}
