% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pseudo.R
\name{pseudo}
\alias{pseudo}
\title{Creates NOT DONE, purpose of function}
\usage{
pseudo(datall, K = 20, c.vec)
}
\arguments{
\item{datall}{A data set with variables X, Y, C, W, G}

\item{K}{The number of folds desired, the default is K=20}

\item{c.vec}{A vector of cutoff values}
}
\value{
NOT DONE
}
\description{
Creates NOT DONE, purpose of function
}
\examples{
pseudo(NOT DONE)

}
