% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{smooth}
\alias{smooth}
\title{Creates NOT DONE, purpose of function}
\usage{
smooth(data, c.vec, K = 20)
}
\arguments{
\item{data}{A data set with variables X, Y, C, W, G, and pseudo outcomes}

\item{c.vec}{A vector of cutoff values}

\item{K}{Number of folds}
}
\value{
NOT DONE: Smoothness parameters
}
\description{
Creates NOT DONE, purpose of function
}
\examples{
smooth(NOT DONE)

}
