% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoother.R
\name{smoother}
\alias{smoother}
\title{Creates NOT DONE, purpose of function}
\usage{
smoother(X, Y, C, c.vec, K)
}
\arguments{
\item{X}{X value}

\item{Y}{Y value}

\item{C}{C value}

\item{c.vec}{A vector of cutoff values}

\item{K}{K folds}
}
\value{
NOT DONE: Smoothness parameters
}
\description{
Creates NOT DONE, purpose of function
}
\examples{
smoother(NOT DONE)

}
