% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{ulbounds}
\alias{ulbounds}
\title{Creates NOT DONE, purpose of function}
\usage{
ulbounds(data, c.vec_initial)
}
\arguments{
\item{data}{A data set with variables X, Y, C, W, G, and pseudo outcomes}

\item{c.vec_initial}{A vector of cutoff values}
}
\value{
NOT DONE: Smoothness parameters
}
\description{
Creates NOT DONE, purpose of function
}
\examples{
ulbounds(NOT DONE)

}
