% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smooth.R
\name{smooth}
\alias{smooth}
\title{Creates NOT DONE, purpose of function}
\usage{
smooth(data_all, c.vec)
}
\arguments{
\item{data_all}{A data set with variables X, Y, C, W, G, and pseudo outcomes}

\item{c.vec}{A vector of cutoff values}
}
\value{
NOT DONE
}
\description{
Creates NOT DONE, purpose of function
}
\examples{
smooth(NOT DONE)

}
